+++ inst/doc/arc-layer.R
@@ -1,29 +1,25 @@
-## ----echo = FALSE--------------------------------------------------------
+## ----echo = FALSE-------------------------------------------------------------
 example <- paste0(
   rprojroot::find_package_root_file(),
   "/inst/examples/deckgl-api-reference/arc-layer.R"
 )
-knitr::read_chunk(example)
+do.call(knitr::read_chunk, list(path = example))
 
-## ----arc-layer, eval = FALSE---------------------------------------------
-#  sample_data <- paste0(
-#    "https://raw.githubusercontent.com/",
-#    "uber-common/deck.gl-data/master/",
-#    "website/bart-segments.json"
-#  )
+## ----arc-layer, eval = FALSE--------------------------------------------------
+#  data("bart_segments")
 #  
 #  properties <- list(
 #    pickable = TRUE,
-#    getStrokeWidth = 12,
-#    getSourcePosition = get_property("from.coordinates"),
-#    getTargetPosition = get_property("to.coordinates"),
+#    getWidth = 12,
+#    getSourcePosition = ~from_lng + from_lat,
+#    getTargetPosition = ~to_lng + to_lat,
 #    getSourceColor = JS("d => [Math.sqrt(d.inbound), 140, 0]"),
 #    getTargetColor = JS("d => [Math.sqrt(d.outbound), 140, 0]"),
-#    getTooltip = JS("object => `${object.from.name} to ${object.to.name}`")
+#    getTooltip = JS("object => `${object.from_name} to ${object.to_name}`")
 #  )
 #  
 #  deck <- deckgl(zoom = 10, pitch = 35) %>%
-#    add_arc_layer(data = sample_data, properties = properties) %>%
+#    add_arc_layer(data = bart_segments, properties = properties) %>%
 #    add_mapbox_basemap()
 #  
 #  if (interactive()) deck