+++ inst/doc/point-cloud-layer.R
@@ -1,11 +1,11 @@
-## ----echo = FALSE--------------------------------------------------------
+## ----echo = FALSE-------------------------------------------------------------
 example <- paste0(
   rprojroot::find_package_root_file(),
   "/inst/examples/deckgl-api-reference/point-cloud-layer.R"
 )
-knitr::read_chunk(example)
+do.call(knitr::read_chunk, list(path = example))
 
-## ----point-cloud-layer, eval = FALSE-------------------------------------
+## ----point-cloud-layer, eval = FALSE------------------------------------------
 #  sample_data <- paste0(
 #    "https://raw.githubusercontent.com/",
 #    "uber-common/deck.gl-data/",
@@ -14,12 +14,12 @@ knitr::read_chunk(example)
 #  
 #  properties <- list(
 #    pickable = TRUE,
-#    coordinateSystem = JS("COORDINATE_SYSTEM.METER_OFFSETS"),
+#    coordinateSystem = JS("deck.COORDINATE_SYSTEM.METER_OFFSETS"),
 #    coordinateOrigin = c(-122.4, 37.74),
-#    radiusPixels = 4,
-#    getPosition = get_property("position"),
-#    getNormal = get_property("normal"),
-#    getColor = get_property("color"),
+#    pointSize = 4,
+#    getPosition = ~position,
+#    getNormal = ~normal,
+#    getColor = ~color,
 #    lightSettings = list(),
 #    getTooltip = JS("object => object.position.join(', ')")
 #  )