+++ inst/doc/scatterplot-layer.R
@@ -1,22 +1,19 @@
-## ----echo = FALSE--------------------------------------------------------
+## ----echo = FALSE-------------------------------------------------------------
 example <- paste0(
   rprojroot::find_package_root_file(),
   "/inst/examples/deckgl-api-reference/scatterplot-layer.R"
 )
-knitr::read_chunk(example)
+do.call(knitr::read_chunk, list(path = example))
 
-## ----scatterplot-layer, eval = FALSE-------------------------------------
-#  bart_stations <- paste0(
-#    "https://raw.githubusercontent.com/",
-#    "uber-common/deck.gl-data/",
-#    "master/website/bart-stations.json"
-#  )
+## ----scatterplot-layer, eval = FALSE------------------------------------------
+#  data("bart_stations")
 #  
 #  properties <- list(
-#    getPosition = get_property("coordinates"),
+#    getPosition = ~lng + lat,
 #    getRadius = JS("data => Math.sqrt(data.exits)"),
 #    radiusScale = 6,
-#    getColor = c(255, 140, 20)
+#    getFillColor = c(255, 140, 20),
+#    getTooltip = ~name
 #  )
 #  
 #  deck <- deckgl(zoom = 10.5, pitch = 35) %>%