+++ inst/doc/text-layer.R
@@ -1,17 +1,19 @@
-## ----echo = FALSE--------------------------------------------------------
+## ----echo = FALSE-------------------------------------------------------------
 example <- paste0(
   rprojroot::find_package_root_file(),
   "/inst/examples/deckgl-api-reference/text-layer.R"
 )
-knitr::read_chunk(example)
+do.call(knitr::read_chunk, list(path = example))
 
-## ----text-layer, eval = FALSE--------------------------------------------
+## ----text-layer, eval = FALSE-------------------------------------------------
+#  data("bart_stations")
+#  
 #  deck <- deckgl(zoom = 10, pitch = 35) %>%
 #    add_text_layer(
 #      data = bart_stations,
 #      pickable = TRUE,
-#      getPosition = get_position("lat", "lng"),
-#      getText = get_property("name"),
+#      getPosition = ~lng + lat,
+#      getText = ~name,
 #      getSize = 15,
 #      getAngle = 0,
 #      getTextAnchor = "middle",