+++ inst/examples/deckgl-api-reference/text-layer.R
@@ -1,10 +1,12 @@
 ## @knitr text-layer
+data("bart_stations")
+
 deck <- deckgl(zoom = 10, pitch = 35) %>%
   add_text_layer(
     data = bart_stations,
     pickable = TRUE,
-    getPosition = get_position("lat", "lng"),
-    getText = get_property("name"),
+    getPosition = ~lng + lat,
+    getText = ~name,
     getSize = 15,
     getAngle = 0,
     getTextAnchor = "middle",