+++ inst/examples/shiny.R
@@ -1,9 +1,13 @@
 ## @knitr shiny-integration
+library(magrittr)
 library(shiny)
 library(deckgl)
 
+.app = reactiveValues(visible = TRUE)
+
 view <- fluidPage(
   h1("deckgl for R"),
+  actionButton("visible", "visible"),
   deckglOutput("deck"),
   tableOutput("selected"),
   style = "font-family: Helvetica, Arial, sans-serif;"
@@ -24,7 +28,16 @@ backend <- function(input, output) {
     object <- info$object
     # print(info)
     print(object$points %>% length())
-    print(object$centroid)
+    print(names(object))
+  })
+
+  observeEvent(input$visible, {
+    print("clicked")
+    .app$visible = ifelse(.app$visible == TRUE, FALSE, TRUE)
+    print(.app$visible)
+    deckgl_proxy("deck") %>%
+      add_hexagon_layer(visible = .app$visible) %>%
+      update_deckgl(it = "works")
   })
 
   df <- eventReactive(input$deck_onclick, {