+++ inst/htmlwidgets/deckgl.js
@@ -4,6 +4,29 @@ var deck = window.deck;
 
   var deckglWidget = window.deckglWidget = {};
 
+  // TODO: Remove, see function below
+  /*
+  deckglWidget.tileLayer = function() {
+    return new deck.TileLayer({
+      pickable: true,
+      opacity: 1,
+      minZoom: 0,
+      maxZoom: 19,
+      renderSubLayers: function(props) {
+        // const tileServer = 'https://c.tile.openstreetmap.org/';
+        const tileServer = "http://a.tile.stamen.com/toner/";
+        const {x, y, z, bbox} = props.tile;
+        const {west, south, east, north} = bbox;
+
+        return new deck.BitmapLayer(props, {
+          image: `${tileServer}/${z}/${x}/${y}.png`,
+          bounds: [west, south, east, north]
+        });
+      }
+    });
+  };
+  */
+
   deckglWidget.colorToRGBArray = function(color) {
     color = color.substring(1); // remove '#'
     return [
@@ -11,6 +34,17 @@ var deck = window.deck;
       parseInt(color.substring(2, 4), 16),
       parseInt(color.substring(4), 16)
     ];
+  };
+
+  deckglWidget.renderMapTiles = function(props) {
+    const tileServer = props.tileServer || "http://a.tile.stamen.com/toner/";
+    const { x, y, z, bbox } = props.tile;
+    const { west, south, east, north } = bbox;
+
+    return new deck.BitmapLayer(props, {
+      image: `${tileServer}/${z}/${x}/${y}.png`,
+      bounds: [ west, south, east, north ]
+    });
   };
 
   var newLayer = function(className, properties) {
@@ -45,6 +79,14 @@ var deck = window.deck;
     }
 
     return new deck[className](properties);
+  };
+
+  var makeDataAccessor = deckglWidget.makeDataAccessor = function(keys) {
+    if (typeof keys === "string") {
+      return data => data[keys];
+    }
+
+    return data => keys.map(key => data[key]);
   };
 
   var initialViewState = function(x) {
@@ -96,13 +138,17 @@ var deck = window.deck;
           // Render the widget
 
           console.log("deck.gl version: " + deck.version);
+          if (typeof mapboxgl !== "undefined") {
+            console.log("mapbox-gl version: " + mapboxgl.version);
+          }
 
           var properties = {
             mapboxApiAccessToken: x.mapboxApiAccessToken || "",
             mapStyle: x.mapStyle || "",
             container: el.id,
             initialViewState: x.initialViewState || initialViewState(x),
-            views: x.views || new MapView(),
+            views: x.views || new deck.MapView(),
+            controller: true,
             layers: []
           };
 
@@ -110,6 +156,7 @@ var deck = window.deck;
 
           deckglWidget.deckgl = deckgl = new deck.DeckGL(properties);
 
+          /*
           deckglWidget.layers = x.layers.map(function(item) {
             if (item.properties.dataframeToD3) {
               item.data = HTMLWidgets.dataframeToD3(item.data);
@@ -119,9 +166,15 @@ var deck = window.deck;
             // console.log(item);
             return newLayer(item.className, item.properties);
           });
+          */
+          deckglWidget.layers = makeLayers(x.layers);
 
           deckgl.setProps({ layers: deckglWidget.layers });
 
+        },
+
+        getDeck: function() {
+          return deckgl;
         },
 
         resize: function(width, height) {
@@ -133,5 +186,58 @@ var deck = window.deck;
       };
     }
   });
+
+  var getWidget = function(id) {
+    var htmlWidgetsObj = HTMLWidgets.find("#" + id);
+    return htmlWidgetsObj.getDeck();
+  };
+
+  if (HTMLWidgets.shinyMode) {
+    Shiny.addCustomMessageHandler('proxythis', function(obj) {
+      var deckgl = getWidget(obj.id);
+
+      // Fix JS properties
+      var layerDefs = obj.x.layers;
+      for (let i = 0; i < layerDefs.length; i++) {
+        var properties = layerDefs[i].properties;
+        for (let key of Object.keys(properties)) {
+          if (typeof properties[key] === "string") {
+            try {
+              properties[key] = eval(properties[key]);
+            } catch(err) { }
+          } // end if
+        } // end for
+      } // end for
+
+      console.log(obj);
+      // console.log(deckgl);
+      var layers = makeLayers(obj.x.layers);
+      deckgl.setProps({ layers: layers });
+      // console.log(deckglWidget.deckgl);
+    });
+  }
+
+  var makeLayers = function(layers) {
+    var l = deckglWidget.layers = layers.map(function(item) {
+      if (item.properties.dataframeToD3) {
+        item.data = HTMLWidgets.dataframeToD3(item.data);
+      }
+
+      item.properties.data = item.data;
+
+      // make data accessors
+      for (let key of Object.keys(item.properties)) {
+        var property = item.properties[key];
+        if (property !== null && typeof property === "object" && property.dataAccessor !== undefined) {
+          console.log(key, "make data accessor");
+          item.properties[key] = makeDataAccessor(property.dataAccessor);
+        }
+      }
+      // -----
+
+      return newLayer(item.className, item.properties);
+    });
+    return l;
+  };
 
 })(); // anonymos end