+++ NEWS.md
@@ -1,10 +1,76 @@
+# deckgl 0.2.5
+
+Renamed
+
+* `icon_definition` to `use_icon_definition`
+* `contour_definition` to `use_contour_definition`
+* `default_icon_properties` to `use_default_icon_properties`
+
+# deckgl 0.2.4
+
+* Switched to `deck.gl.js` v8.0.16
+* Switched to `mapbox-gl.js` v1.8.0
+* Refactored dependencies (uses `deps.yaml` now)
+* Added `use_carto_style` and `add_basemap` using a basemap from carto as default
+* Added sample data for column and grid cell layer
+* Changed parameter order in `add_mapbox_basemap`, so that `style` is the first parameter now
+
+# deckgl 0.2.3
+
+* Switched to `deck.gl.js` v8.0.15
+* `add_heatmap_layer`
+* `add_grid_cell_layer`
+
+# deckgl 0.2.2
+
+* Added `deckgl_proxy` and `update_deckgl` to update layers in _shiny_ mode
+
+# deckgl 0.2.1
+
+* Switched to `deck.gl.js` v7.1.7
+* Added support for `sf` objects as `data` parameter
+* Added _formula_ syntax for data accessors
+* Added more examples (see  folder `sf` on how to pass `sf` objects to the layers)
+
+```r
+properties <- list(
+  getPosition = JS("d => [d.lng, d.lat]")
+  getRadius = JS("d => d.radius")
+)
+
+# equals
+
+properties <- list(
+  getPosition = ~lng + lat,
+  getRadius = ~radius
+)
+```
+
+# deckgl 0.2.0
+
+* Switched to `deck.gl.js` v7.1.6
+* `add_column_layer`
+* `add_great_circle_layer`
+* `add_bitmap_layer`
+* `add_h3_cluster_layer`
+* `add_h3_hexagon_layer`
+* `add_raster_tile_layer`
+
+# deckgl 0.1.9
+
+* Switched to `deck.gl.js` v6.3.0
+* Switched to `mapbox-gl.js` v0.51.0
+* Updated api example of `ScreenGridLayer`
+* Added `HexagonLayer` example of _road-safety-in-uk_
+
 # deckgl 0.1.8
 
-* Switched to `deckgl.js` 6.2.4
+* Switched to `deck.gl.js` v6.2.4
+* cran release
 
 # deckgl 0.1.7
 
-* Switched to `deckgl.js` 6.2.3
+* Switched to `deck.gl.js` v6.2.3
 * Added examples to r-docs for all `add_*_layer` functions
 * Added helper functions to run api-examples, not exported yet:
 
@@ -16,7 +82,7 @@ deck
 
 # deckgl 0.1.6
 
-* Switched to `deckgl.js` 6.2.2
+* Switched to `deck.gl.js` v6.2.2
 * Added more examples
 * Added `...` parameter to `deckgl` to pass optional parameters to the `deck` instance:
 
@@ -38,7 +104,7 @@ deckgl() %>%
 
 # deckgl 0.1.5
 
-* Switched to `deckgl.js` 6.1.1
+* Switched to `deck.gl.js` 6.1.1
 * Added helper functions `get_property` and `get_color_to_rgb_array`:
 
 ```r
@@ -59,7 +125,7 @@ properties <- list(
 
 # deckgl 0.1.4
 
-* Switched to `deckgl.js` 6.1.0
+* Switched to `deck.gl.js` v6.1.0
 * Added missing `add_*_layer` functions for all core layers including `add_contour_layer` (new in version 6.1.0)
 * Added helper functions `encode_icon_atlas`, `icon_definition` and `default_icon_properties` to set icons for the icon layer: